+++ frontend/package.json
@@ -11,10 +11,10 @@
     "bootstrap": "5.1",
     "cra-template": "1.2.0",
     "react": "^19.0.0",
-    "react-cookie": "^4.1.1",
+    "react-cookie": "^7.2.2",
     "react-dom": "^19.0.0",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "5.0.1",
+    "react-router-dom": "^6.28.2",
+    "react-scripts": "^5.0.1",
     "reactstrap": "^8.10.0"
   },
   "scripts": {
@@ -43,6 +43,7 @@
   },
   "devDependencies": {
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@types/react-router-dom": "^5.3.3",
     "web-vitals": "^4.2.4"
   }
 }
