+++ frontend/src/components/search/SearchBar.css
